src/api/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
psycopg2-binary==2.9.10
requests==2.32.3
redis==5.2.1
rq==1.16.2